% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.ria.test}
\alias{tidy.ria.test}
\title{Tidy a ria.test object}
\usage{
\method{tidy}{ria.test}(x, ...)
}
\arguments{
\item{x}{A `ria.test` object produced by a call to [ria.test::ria.test()].}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing information about the model.
}
\description{
Tidy a ria.test object
}
\examples{
\donttest{
if (require("mma") && torch::torch_is_installed()) {
	library(mma)
	data(weight_behavior)

	weight_behavior <- na.omit(weight_behavior)

	res <- ria.test(
		data = weight_behavior,
		trt = "sports",
		outcome = "bmi",
		covar = c("age", "sex", "tvhours"),
		mediators = c("exercises", "overweigh"),
		moc = "snack",
		d0 = \(data, trt) factor(rep(1, nrow(data)), levels = c("1", "2")),
		d1 = \(data, trt) factor(rep(2, nrow(data)), levels = c("1", "2")),
		learners = c("mean", "glm"),
		nn_module = sequential_module(),
		control = ria.test.control(crossfit_folds = 1L, zprime_folds = 5L, epochs = 10L)
	)

	print(res)
	tidy(res)
}
}
}
